app/requirements.txt
Flask==2.0.2
google-cloud-firestore==2.4.0
google-cloud-secret-manager==2.9.2
google-cloud-storage==2.2.1
# imgkit==1.1.2
